requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.10.0
django-cors-headers==4.3.1
django-filter==23.3
django-redis>=5.4.0
djangorestframework==3.14.0
idna==3.10
oauthlib==3.3.1
Pillow==10.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0

